agent/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
openai==1.50.0
anthropic==0.34.0
google-generativeai==0.8.0
python-dotenv==1.0.1
httpx==0.27.0
